sicei/requirements.txt
Django==3.2.6
djangorestframework==3.12.4
drf-yasg==1.20.0
psycopg2==2.9.1
django-environ==0.4.5
gunicorn==20.1.0